requirements.txt
pyrogram==1.4.16
tgcrypto
youtube-dl
aiohttp
aiofiles
ffmpeg-python
pillow
hachoir
motor
dnspython
python-dotenv
telethon
requests
